Rename `PackageInternal` annotation to 'PackageProtected'
