--- std/package-internal-lint/src/main/resources/META-INF/com.android.tools.lint.client.api.IssueRegistry
@@ -1,1 +1,0 @@
-br.com.orcinus.orca.std.packageinternal.lint.PackageInternalIssueRegistry
