+++ std/package-protected-lint/src/main/resources/META-INF/com.android.tools.lint.client.api.IssueRegistry
@@ -1,0 +1,1 @@
+br.com.orcinus.orca.std.packageprotected.lint.PackageProtectedIssueRegistry
