+++ std/package-protected-lint/src/test/java/br/com/orcinus/orca/std/packageprotected/lint/detection/PackageProtectedDetectorTests.kt
@@ -1,0 +1,191 @@
+/*
+ * Copyright © 2024 Orcinus
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+ * the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program. If
+ * not, see https://www.gnu.org/licenses.
+ */
+
+package br.com.orcinus.orca.std.packageprotected.lint.detection
+
+import com.android.tools.lint.checks.infrastructure.TestFiles
+import com.android.tools.lint.checks.infrastructure.TestLintTask
+import kotlin.test.Test
+
+internal class PackageProtectedDetectorTests {
+  private val lintTask =
+    TestLintTask.lint().allowMissingSdk().issues(PackageProtectedDetector.issue)
+
+  @Test
+  fun reportsOnReferenceFromOutsidePackageToClassMarkedAsPackageProtectedFromCustomAnnotation() {
+    lintTask
+      .files(
+        TestFiles.packageProtectedAnnotation,
+        TestFiles.packageProtectedAnnotatedAnnotation,
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.std.packageprotected.lint.detection.test
+
+              import br.com.orcinus.orca.std.packageprotected.PackageProtected
+
+              @PackageProtectedApi
+              class Api
+            """
+          )
+          .indented(),
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.app
+
+              import br.com.orcinus.orca.std.packageprotected.lint.detection.test.Api
+
+              private object Consumer {
+                init {
+                  Api()
+                }
+              }
+            """
+          )
+          .indented()
+      )
+      .run()
+      .expect(
+        """
+          src/br/com/orcinus/orca/app/Consumer.kt:7: Error: ${PackageProtectedDetector.MESSAGE} [${PackageProtectedDetector.issue.id}]
+              Api()
+              ~~~~~
+          1 errors, 0 warnings
+        """
+      )
+  }
+
+  @Test
+  fun reportsOnReferenceFromOutsidePackageToPackageProtectedClass() {
+    lintTask
+      .files(
+        TestFiles.packageProtectedAnnotation,
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.std.packageprotected.lint.detection.test
+
+              import br.com.orcinus.orca.std.packageprotected.PackageProtected
+
+              @PackageProtected
+              class Api
+            """
+          )
+          .indented(),
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.app
+
+              import br.com.orcinus.orca.std.packageprotected.lint.detection.test.Api
+
+              private object Consumer {
+                init {
+                  Api()
+                }
+              }
+            """
+          )
+          .indented()
+      )
+      .run()
+      .expect(
+        """
+          src/br/com/orcinus/orca/app/Consumer.kt:7: Error: ${PackageProtectedDetector.MESSAGE} [${PackageProtectedDetector.issue.id}]
+              Api()
+              ~~~~~
+          1 errors, 0 warnings
+        """
+      )
+  }
+
+  @Test
+  fun reportsOnReferenceFromOutsidePackageToClassConstructorMarkedAsPackageProtectedFromCustomAnnotation() {
+    lintTask
+      .files(
+        TestFiles.packageProtectedAnnotation,
+        TestFiles.packageProtectedAnnotatedAnnotation,
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.std.packageprotected.lint.detection.test
+
+              class Api @PackageProtectedApi constructor()
+            """
+          )
+          .indented(),
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.app
+
+              import br.com.orcinus.orca.std.packageprotected.lint.detection.test.Api
+
+              private object Consumer {
+                init {
+                  Api()
+                }
+              }
+            """
+          )
+          .indented()
+      )
+      .run()
+      .expect(
+        """
+          src/br/com/orcinus/orca/app/Consumer.kt:7: Error: ${PackageProtectedDetector.MESSAGE} [${PackageProtectedDetector.issue.id}]
+              Api()
+              ~~~~~
+          1 errors, 0 warnings
+        """
+      )
+  }
+
+  @Test
+  fun reportsOnReferenceFromOutsidePackageToPackageProtectedClassConstructor() {
+    lintTask
+      .files(
+        TestFiles.packageProtectedAnnotation,
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.std.packageprotected.lint.detection.test
+
+              import br.com.orcinus.orca.std.packageprotected.PackageProtected
+
+              class Api @PackageProtected constructor()
+            """
+          )
+          .indented(),
+        TestFiles.kotlin(
+            """
+              package br.com.orcinus.orca.app
+
+              import br.com.orcinus.orca.std.packageprotected.lint.detection.test.Api
+
+              private object Consumer {
+                init {
+                  Api()
+                }
+              }
+            """
+          )
+          .indented()
+      )
+      .run()
+      .expect(
+        """
+          src/br/com/orcinus/orca/app/Consumer.kt:7: Error: ${PackageProtectedDetector.MESSAGE} [${PackageProtectedDetector.issue.id}]
+              Api()
+              ~~~~~
+          1 errors, 0 warnings
+        """
+      )
+  }
+}
