+++ std/package-protected/build.gradle.kts
